analyzer: add determinism for various output modes
While the choice of call path the tool displays to demonstrate why a function has a particular capability is somewhat arbitrary, it's useful if the choice is consistent across runs so that outputs can be compared more easily.

To achieve this, we add sorts in various places, and improve the existing sorts for graph nodes and edges so there are no spurious ties between items.