analyzer: fix some cases of incorrect AST rewriting
The rewriteCallsToSort function uses astutil.Apply to modify the AST.  These
modifications replace single statements with multiple statements, but this is
not valid in all parts of the syntax tree, e.g. the init part of a for
statement.  So here we check the context of a statement before rewriting it.