analyzer: some changes to call graph generation
Change the outputCall parameter of CapabilityGraph to include, alongside
the callgraph edge, the BFS tree that led from the query to the edge's
caller, and from the edge's callee to a capability.  Users of
CapabilityGraph can use this information to reconstruct an example call
path through the graph that includes the given edge.

Add a type definition for the BFS state, since it is repeated
frequently.

Add types for the two callback function parameters of CapabilityGraph.

Make the search order for graph output deterministic, by considering
edges in sorted order when visiting a node.  This is similar to what is
done for other output modes.

Add a byCallee type, similar to the byCaller type, for sorting callgraph
edges by their callee.