analyzer: add "intermediate" granularity
This granularity finds which packages exist along a path in the callgraph from
a function in a queried package to a function with a capability.

This allows users to see when a change in their dependencies results in a
capability being reached in a path via a new package, even if the top-level
function already had that capability through another path.