cmd/capslock: add a flag which disables use of temporary modules
When the requested packages are not dependencies of the current
workspace, or all the requested packages use wildcards and none of them
match any dependencies, capslock will `go get` the packages in a
temporary module in order to analyze them.

This change adds a flag -force_local_module which disables this
behavior, so that the user can restrict the analysis to dependencies
of the current module and return an error if this fails.