Introduce fine-grained capabilities with string representation
This change refactors the capability model to allow for more specific
and descriptive capability names.  Previously, capabilities were
represented by an enum that was not very granular.  For example,
`MODIFY_SYSTEM_STATE` covered a wide range of operations, from changing
environment variables to altering system logging.

This commit introduces a new string-based representation for
capabilities, allowing for finer-grained distinctions such as
`MODIFY_SYSTEM_STATE/ENV` and `MODIFY_SYSTEM_STATE/LOGGING`.  It also
makes it easier to specify custom capabilities, and to add new
capabilities that make sense for other languages.

This change maintains some backwards compatibility by still reading and
writing the enum field.  When writing, enums are calculated by looking
for either exact matches to the string, or matches of a prefix of the
string before the first '/' character, which maps fine-grained
capabilities back to the coarser enum.

This commit changes the default function classifier to use finer-grained
capabilities for MODIFY_SYSTEM_STATE, but leaves others alone for now.